scrapy-crawlers/requirements.txt
Scrapy==1.5.0
fake-useragent==0.1.10
beautifulsoup4==4.6.0
elasticsearch==6.3.0
python_dateutil==2.7.3